[refactor] Move TypedConstants to taichi/ir/type (#2211)
* [refactor] Move TypedConstants to taichi/ir/type

* [skip ci] enforce code format

* rm

* sort
