+++ taichi/ir/frontend_ir.h
@@ -10,61 +10,6 @@
 
 TLANG_NAMESPACE_BEGIN
 
-class ASTBuilder;
-
-ASTBuilder &current_ast_builder();
-
-class FrontendContext {
- private:
-  std::unique_ptr<ASTBuilder> current_builder;
-  std::unique_ptr<Block> root_node;
-
- public:
-  FrontendContext();
-
-  ASTBuilder &builder() {
-    return *current_builder;
-  }
-
-  IRNode *root();
-
-  std::unique_ptr<Block> get_root() {
-    return std::move(root_node);
-  }
-};
-
-extern std::unique_ptr<FrontendContext> context;
-
-class ASTBuilder {
- private:
-  std::vector<Block *> stack;
-
- public:
-  ASTBuilder(Block *initial) {
-    stack.push_back(initial);
-  }
-
-  void insert(std::unique_ptr<Stmt> &&stmt, int location = -1);
-
-  struct ScopeGuard {
-    ASTBuilder *builder;
-    Block *list;
-    ScopeGuard(ASTBuilder *builder, Block *list)
-        : builder(builder), list(list) {
-      builder->stack.push_back(list);
-    }
-
-    ~ScopeGuard() {
-      builder->stack.pop_back();
-    }
-  };
-
-  std::unique_ptr<ScopeGuard> create_scope(std::unique_ptr<Block> &list);
-  Block *current_block();
-  Stmt *get_last_stmt();
-  void stop_gradient(SNode *);
-};
-
 // Frontend Statements
 
 class FrontendAllocaStmt : public Stmt {
@@ -630,5 +575,58 @@ class ExternalTensorShapeAlongAxisExpression : public Expression {
 
   void flatten(FlattenContext *ctx) override;
 };
+
+class ASTBuilder {
+ private:
+  std::vector<Block *> stack;
+
+ public:
+  ASTBuilder(Block *initial) {
+    stack.push_back(initial);
+  }
+
+  void insert(std::unique_ptr<Stmt> &&stmt, int location = -1);
+
+  struct ScopeGuard {
+    ASTBuilder *builder;
+    Block *list;
+    ScopeGuard(ASTBuilder *builder, Block *list)
+        : builder(builder), list(list) {
+      builder->stack.push_back(list);
+    }
+
+    ~ScopeGuard() {
+      builder->stack.pop_back();
+    }
+  };
+
+  std::unique_ptr<ScopeGuard> create_scope(std::unique_ptr<Block> &list);
+  Block *current_block();
+  Stmt *get_last_stmt();
+  void stop_gradient(SNode *);
+};
+
+ASTBuilder &current_ast_builder();
+
+class FrontendContext {
+ private:
+  std::unique_ptr<ASTBuilder> current_builder;
+  std::unique_ptr<Block> root_node;
+
+ public:
+  FrontendContext();
+
+  ASTBuilder &builder() {
+    return *current_builder;
+  }
+
+  IRNode *root();
+
+  std::unique_ptr<Block> get_root() {
+    return std::move(root_node);
+  }
+};
+
+extern std::unique_ptr<FrontendContext> context;
 
 TLANG_NAMESPACE_END
