+++ taichi/ir/statements.h
@@ -839,17 +839,19 @@ class OffloadedStmt : public Stmt {
   using TaskType = OffloadedTaskType;
 
   TaskType task_type;
-  SNode *snode;
-  std::size_t begin_offset;
-  std::size_t end_offset;
-  bool const_begin, const_end;
-  int32 begin_value, end_value;
-  int step;
+  Arch device;
+  SNode *snode{nullptr};
+  std::size_t begin_offset{0};
+  std::size_t end_offset{0};
+  bool const_begin{false};
+  bool const_end{false};
+  int32 begin_value{0};
+  int32 end_value{0};
+  int step{0};
   int grid_dim{1};
   int block_dim{1};
-  bool reversed;
-  int num_cpu_threads;
-  Arch device;
+  bool reversed{false};
+  int num_cpu_threads{1};
 
   std::vector<int> index_offsets;
 
@@ -862,9 +864,7 @@ class OffloadedStmt : public Stmt {
   std::size_t bls_size{0};
   MemoryAccessOptions mem_access_opt;
 
-  OffloadedStmt(TaskType task_type);
-
-  OffloadedStmt(TaskType task_type, SNode *snode);
+  OffloadedStmt(TaskType task_type, Arch arch);
 
   std::string task_name() const;
 
@@ -882,8 +882,9 @@ class OffloadedStmt : public Stmt {
 
   void all_blocks_accept(IRVisitor *visitor);
 
-  TI_STMT_DEF_FIELDS(ret_type,
+  TI_STMT_DEF_FIELDS(ret_type /*inherited from Stmt*/,
                      task_type,
+                     device,
                      snode,
                      begin_offset,
                      end_offset,
@@ -896,7 +897,6 @@ class OffloadedStmt : public Stmt {
                      block_dim,
                      reversed,
                      num_cpu_threads,
-                     device,
                      index_offsets,
                      mem_access_opt);
   TI_DEFINE_ACCEPT
