+++ taichi/program/ir_node_extended_impl.cpp
@@ -3,6 +3,10 @@
 #include "taichi/program/program.h"
 #include "taichi/program/kernel.h"
 
+// TODO(#2196): Part of the implementation of IRNode has to be placed here, in
+// order to use files under "taichi/program". Ideally, we should:
+// 1. Have a dedicated IR compile config that lives under "taichi/ir"
+// 2. Just don't hook IRNode with a Kernel
 namespace taichi {
 namespace lang {
